cleaned up style files and merged into one file for navigation for ease of use and readability
