+++ src/App.js
@@ -1,16 +1,15 @@
 import { Route } from "react-router-dom";
 import "./App.css";
+import ImageSlider from "./components/ImageSlider";
 import Navbar from "./components/Navbar";
 import GlobalStyle from "./styles/global";
 
 function App() {
     return (
         <div className="App">
-          <GlobalStyle/>
+            <GlobalStyle />
             <Navbar />
-            <div>
-                <h1>Design 3dwrgergrgr</h1>
-            </div>
+            <ImageSlider />
         </div>
     );
 }
