+++ src/components/ImageSlider.js
@@ -1,0 +1,15 @@
+import React from "react";
+import styled, { css } from "styled-components/macro";
+import {HeroSection} from "../styles/imageSlider/Hero"
+
+
+
+const ImageSlider = () => {
+    return (
+        <HeroSection>
+            <h1>Image</h1>
+        </HeroSection>
+    );
+};
+
+export default ImageSlider;
