+++ src/styles/imageSlider/Hero.js
@@ -1,0 +1,9 @@
+import styled from "styled-components"
+
+
+export const HeroSection = styled.section`
+    height: 100vh;
+    max-height: 1100px;
+    position: relative;
+    overflow: hidden;
+`;
