--- src/styles/navigation/logo.js
@@ -1,13 +1,0 @@
-import styled from "styled-components";
-import { Link } from "react-router-dom";
-
-export const Logo = styled(Link)`
-    font-style: italic;
-    font-weight: bold;
-    color: #fff;
-    display: flex;
-    align-items: center;
-    padding: 0 1rem;
-    height: 100%;
-    cursor: pointer;
-`;
