--- src/styles/navigation/navMenu.js
@@ -1,12 +1,0 @@
-import styled from "styled-components";
-
-export const NavMenu = styled.div`
-    display: flex;
-    align-items: center;
-    margin-right: -48px;
-
-
-    @media screen and (max-width: 768px){
-        display:none;
-    }
-`;
